requirements.txt
alembic==1.10.4
anyio==3.6.2
asyncpg==0.27.0
attrs==23.1.0
bcrypt==4.0.1
boto3==1.26.129
botocore==1.29.129
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
databases==0.7.0
distro==1.8.0
dnspython==2.3.0
email-validator==2.1.0.post1
exception==0.1.0
fastapi==0.95.1
FastAPI-SQLAlchemy==0.2.1
fonttools==4.39.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httptools==0.5.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
jmespath==1.0.1
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mirakuru==2.5.2
motor==3.1.2
numpy==1.24.3
outcome==1.3.0
packaging==23.1
pandas==2.0.1
passlib==1.7.4
Pillow==9.5.0
pluggy==1.3.0
port-for==0.7.2
psutil==5.9.6
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.6
pydantic==1.10.7
PyJWT==2.8.0
pymongo==4.3.3
pyparsing==3.0.9
pytest==7.4.2
pytest-dependency==0.5.1
pytest-postgresql==5.0.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.48
starlette==0.26.1
trio==0.22.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
redis==5.0.1
Jinja2==3.1.2
GeoAlchemy2==0.14.6
httpx==0.25.0
shapely==2.0.3